fix: use 'integer' type for JSON Structure to allow native JSON numbers
The int64 type in JSON Structure expects string-encoded integers for
JSON safety with large numbers. Use 'integer' instead which accepts
native JSON number values.

Add test_jstruct_instance_validates_with_sdk to verify input JSON
instances validate against generated schemas.

+++ avrotize/schema_inference.py
@@ -301,8 +301,10 @@ class JsonStructureSchemaInferrer(SchemaInferrer):
     """Infers JSON Structure schemas from JSON and XML data."""
 
     # JSON Structure primitive type mapping
+    # Use 'integer' for general integers (accepts native JSON numbers)
+    # int64/uint64 etc. are string-encoded for JSON safety with large numbers
     PYTHON_TO_JSTRUCT_TYPES = {
-        int: "int64",  # Use int64 for safety with large integers
+        int: "integer",
         float: "double",
         str: "string",
         bool: "boolean",
